Fix numeric columns returning as strings from PostgreSQL
Change all decimal columns to float so the pg driver returns JS numbers
instead of strings. decimal columns are returned as strings by the node
postgres driver to preserve precision, causing string concatenation
instead of addition in frontend aggregations (e.g. Portfolio Value
showing 158,250M instead of $530K).

+++ src/entities/case/case.entity.ts
@@ -53,22 +53,22 @@ export class CaseEntity extends BaseEntity {
   @Column({ nullable: true })
   assigned_attorney: string;
 
-  @Column({ nullable: true, type: 'decimal' })
+  @Column({ nullable: true, type: 'float' })
   estimated_value_low: number;
 
-  @Column({ nullable: true, type: 'decimal' })
+  @Column({ nullable: true, type: 'float' })
   estimated_value_high: number;
 
-  @Column({ nullable: true, type: 'decimal' })
+  @Column({ nullable: true, type: 'float' })
   case_strength_score: number;
 
-  @Column({ nullable: true, type: 'decimal' })
+  @Column({ nullable: true, type: 'float' })
   credibility_score: number;
 
-  @Column({ nullable: true, type: 'decimal' })
+  @Column({ nullable: true, type: 'float' })
   settlement_probability: number;
 
-  @Column({ nullable: true, type: 'decimal' })
+  @Column({ nullable: true, type: 'float' })
   fraud_score: number;
 
   @Column({ nullable: true, type: 'text' })
